During GC write victim pages to different sections.
Writing the victim pages to the same section as that of
the current sections accepting user I/O will write
pages with varied modification times to the segments.
The average segment type may not be close to the page
modification time and cost based cleaning will not
work.

Writing GC IO to a new segment.

This needs change to mkfs as well
As you need to add current segments in the
checkpointing area.

Needs testing. May not be complete.

TODO:
1) defrag data - where do we write this?
It should definitely not go in the current segment.
So think of this as well.

Covers the writeback path.
-----

Bug fixing in the __write_node_path()

improved the classification for segment selection.
Fixed a bug where the wrong segment was getting
allocated from in case of retrial

Signed-off-by: Surbhi Palande <csurbhi@gmail.com>