Calculate the write amplification
maintain two counters - one for GC writes and
another for application writes. Increment respectively
before writing out the page. Print when the counters
are full (this is racy and in the worst case loose
one or two writes - but this should not affect
the overall amplification much - so for now this is ok)
(ideally since there are two atomic operations, we want
a lock before accessing them)

Print the counters on unmount as well.

Signed-off-by: Surbhi Palande <csurbhi@gmail.com>