Select more zones of the same type.
This is not necessary when the disk is sparsely full.
But as the disk gets full and space is reused, we will
need a complete patch.

When we select a victim of the requested type, here are the problems:
a) You may not find any more segments of the same type (should be always
possible to end in this case - even without this patch)
b) You may end up with the same seg. So add in a unique manner and when
segment is the same -that indicates you have come back. So stop
searching any more. This wont be possible if you are always only
searching 8GB

Signed-off-by: Surbhi Palande <csurbhi@gmail.com>