Small error
Had not marked bg gc page as gc page
this could affect the write amplification metrics

Not related to the NULL pointer bug.
Do not see the bug in the do_garbage_collect
and family of functions.

Bug is seen on a page that is not valid
but sync inodes tries to write to it.

Signed-off-by: Surbhi Palande <csurbhi@gmail.com>