The blk_plug bug seems to be the following:
When the GC thread sleeps, scheduler flushed out any
pending I/Os that the thread has in
tsk->plug.

tsk->plug becomes NULL when we call blk_finish_plug()
Thus a sleeping thread has nothing to flush when
blk_finish_plug() is called.

Now our plug in all functions is a stack variable
that vanishes when we return from the function.
Thus when scheduler tries to flush the plugged I/Os,
plug is not NULL and is an invalid address and
thus this becomes a fault.

So, this patch makes it simple to start
and finish plugging - by putting it in one
place - rather than making it conditional and
in a loop

Signed-off-by: Surbhi Palande <csurbhi@gmail.com>